# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Yeungjin-Global",
        "description": "YJG API Documentation",
        "version": "0.1"
    },
    "paths": {
...
        "/api/notice": {
            "get": {
                "tags": [
                    "공지사항"
                ],
                "summary": "공지사항 검색",
                "description": "body로 보낸 값에 맞는 공지사항을 검색하는 기능입니다. 검색할 값을 보내지 않으면 전체 공지사항을 반환합니다.",
                "operationId": "ed717a5273b1f0fe69b37754d9f9d4ae",
                "requestBody": {
                    "description": "페이지, 공지사항의 태그(admin, salon, restaurant, bus), 제목으로 검색합니다.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "page": {
                                        "description": "페이지",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tag": {
                                        "description": "태그",
                                        "type": "string",
                                        "example": "Bus"
                                    },
                                    "title": {
                                        "description": "검색어",
                                        "type": "string",
                                        "example": "엄준식"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "ValidationException"
                    },
                    "500": {
                        "description": "ServerError"
                    }
                }
            },
            "post": {
                "tags": [
                    "공지사항"
                ],
                "summary": "공지사항 작성(관리자)",
                "description": "공지사항 작성에 사용됩니다. 사진은 이미지 파일 그대로 보내면 됩니다.",
                "operationId": "af5200b6f4c7031334269d8f6162d2dd",
                "requestBody": {
                    "description": "공지사항에 들어갈 내용 및 사진",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "제목",
                                        "type": "string",
                                        "example": "제목입니다."
                                    },
                                    "content": {
                                        "description": "내용",
                                        "type": "string",
                                        "example": "내용입니다."
                                    },
                                    "tag": {
                                        "description": "태그",
                                        "type": "string",
                                        "example": "행정"
                                    },
                                    "images": {
                                        "type": "array",
                                        "items": {
                                            "example": "file"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "ValidationException"
                    },
                    "500": {
                        "description": "ServerError"
                    }
                }
            }
        },
        "/api/notice/recent": {
            "get": {
                "tags": [
                    "공지사항"
                ],
                "summary": "최근 3건의 공지사항",
                "description": "Query string과 일치하는 태그를 가진 최신 공지사항 3개를 반환합니다.",
                "operationId": "cbcdfc23b6bb7395b90a1cf7f8687b4a",
                "parameters": [
                    {
                        "name": "tag",
                        "in": "query",
                        "description": "찾을 공지사항의 태그",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "ValidationException"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/notice/recent/urgent": {
            "get": {
                "tags": [
                    "공지사항"
                ],
                "summary": "최근 긴급 공지사항",
                "description": "bus, admis 태그를 가진 최신 긴급 공지사항 1개를 반환합니다.",
                "operationId": "2e9d9b83c12159317b5baffad8435d7a",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "ValidationException"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/notice/{id}": {
            "get": {
                "tags": [
                    "공지사항"
                ],
                "summary": "공지사항 검색",
                "description": "아이디와 일치하는 공지사항을 반환합니다.",
                "operationId": "1c64bdae7a0193c179e08e51221a81e5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "찾을 공지사항의 아이디",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "ModelNotFoundException"
                    },
                    "500": {
                        "description": "ServerError"
                    }
                }
            },
            "delete": {
                "tags": [
                    "공지사항"
                ],
                "summary": "공지사항 삭제(관리자)",
                "description": "공지사항 삭제 시 사용됩니다.",
                "operationId": "fb8e594d035f144948c965e87861738b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "삭제할 공지사항의 아이디",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            },
            "patch": {
                "tags": [
                    "공지사항"
                ],
                "summary": "공지사항 수정(관리자)",
                "description": "공지사항을 수정할 때 사용됩니다. 수정할 값만 request body에 전송하면 됩니다.\n     *                  삭제할 이미지의 경우에는 delete_images 배열에 이미지의 아이디 값을 문자열로,\n     *                  추가할 이미지는 기존 images 배열에 파일로 추가하면 됩니다.",
                "operationId": "e441ee7d745bd097bd3a7e161ba3b0f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "수정할 공지사항의 아이디",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "수정할 글 내용",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "제목",
                                        "type": "string",
                                        "example": "제목입니다."
                                    },
                                    "content": {
                                        "description": "내용",
                                        "type": "string",
                                        "example": "내용입니다."
                                    },
                                    "urgent": {
                                        "description": "긴급",
                                        "type": "boolean",
                                        "example": "긴급 여부"
                                    },
                                    "tag": {
                                        "description": "태그",
                                        "type": "string",
                                        "example": "행정"
                                    },
                                    "images": {
                                        "type": "array",
                                        "items": {
                                            "example": "string"
                                        }
                                    },
                                    "delete_images": {
                                        "type": "array",
                                        "items": {
                                            "example": "file"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "ModelNotFoundException"
                    },
                    "422": {
                        "description": "ValidationException"
                    },
                    "500": {
                        "description": "ServerError"
                    }
                }
            }
        },
...
    }
}